Switching classes from the Controller package to Model
